refactor: ParserTest -> CarsTest Migration
